requirements.txt
# amqp==5.1.1
# asgiref==3.7.2
# asn1crypto==1.5.1
# billiard==4.1.0
# blinker==1.6.2
# cachetools==5.3.1
# celery==5.3.4
# certifi==2023.7.22
# cffi==1.16.0
# charset-normalizer==2.0.12
# click==8.1.7
# click-didyoumean==0.3.0
# click-plugins==1.1.1
# click-repl==0.3.0
# cloudinary==1.34.0
# colorama==0.4.6
# cryptography==41.0.4
Django==3.2.8
jango-cors-headers==4.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
# drf-yasg==1.21.7
# Faker==19.11.0
# Flask==2.3.3
# google-api-core==2.12.0
# google-auth==2.23.2
# google-cloud-speech==2.21.0
# googleapis-common-protos==1.60.0
# grpcio==1.59.0
# grpcio-status==1.59.0
gunicorn==21.2.0
# idna==3.4
# inflection==0.5.1
# itsdangerous==2.1.2
# Jinja2==3.1.2
# kombu==5.3.2
# MarkupSafe==2.1.3
mysqlclient==2.2.0
# numpy==1.26.0
# oscrypto==1.3.0
# packaging==23.2
# pandas==2.1.1
# pika==1.3.2
# Pillow==10.0.1
# prompt-toolkit==3.0.39
# proto-plus==1.22.3
# protobuf==4.24.3
# psycopg2-binary==2.9.9
# pyasn1==0.5.0
# pyasn1-modules==0.3.0
# pycparser==2.21
# pyHanko==0.20.1
# pyhanko-certvalidator==0.24.1
# PyJWT==2.8.0
# pypdf==3.16.4
# pypng==0.20220715.0
# python-bidi==0.4.2
# python-dateutil==2.8.2
# python-decouple==3.8
# python-dotenv==1.0.0
# pytz==2023.3.post1
# PyYAML==6.0.1
# qrcode==7.4.2
# redis==4.6.0
# reportlab==4.0.6
# requests==2.26.0
# rsa==4.9
# sentry-sdk==1.32.0
# six==1.16.0
# sqlparse==0.4.4
# typing_extensions==4.8.0
# tzdata==2023.3
# tzlocal==5.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.16
# vine==5.0.0
# wcwidth==0.2.8
# Werkzeug==2.3.7
# whitenoise==6.6.0
# XlsxWriter==3.1.7
